requirements.txt
boto3
fastparquet
google-cloud-bigquery
intake
intake-parquet
requests
s3fs
